+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [2023.2.6]
 
+### `OCC` enhancements
+- Added code completion and references of Bean properties in the Level Mapping tags [#578](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/578)  
+
 ### `Groovy Script` enhancements
 - Introduced actions toolbar for `.groovy` files [#564](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/564)
 - Added `Open Groovy Script` action